django_app/requirements.txt
appdirs==1.4.4
arandr==0.1.10
attrs==22.2.0
autocommand==2.2.2
backports.entry-points-selectable==1.1.1
btrfsutil==6.1.2
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
colorama==0.4.6
commonmark==0.9.1
contextlib2==21.6.0
cryptography==38.0.4
cupshelpers==1.0
Cython==0.29.33
distlib==0.3.6
distro==1.8.0
djangorestframework-camel-case==1.3.0
djangorestframework-yaml==2.0.0
dnspython==2.2.1
docutils==0.19
filelock==3.4.0
future==0.18.2
greenlet==2.0.1
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
inflect==6.0.2
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
lit==14.0.6.dev0
louis==3.24.0
Markdown==3.4.1
more-itertools==9.0.0
msgpack==1.0.4
netsnmp-python==1.0a1
nftables==0.1
nspektr==0.4.0
ordered-set==4.1.0
packaging==21.3
pbr==5.8.0
pep517==0.13.0
platformdirs==2.6.2
ply==3.11
progress==1.6
pycairo==1.23.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.4
Pygments==2.14.0
PyGObject==3.42.2
pynvim==0.4.3
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyYAML==6.0
ranger-fm==1.9.3
Reflector==2021.11.20.2.41.3
requests==2.28.1
resolvelib==0.9.0
retrying==1.3.3
rich==13.0.1
six==1.16.0
stevedore==3.5.0
TBB==0.2
team==1.0
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.12.22
typing_extensions==4.4.0
ufw==0.36.1
urllib3==1.26.12
validate-pyproject==0.10.1
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webencodings==0.5.1
zipp==3.11.0
